+++ src/js/src/views/Logs.vue
@@ -301,7 +301,8 @@
           .then((response) => {
             this.logs = response.data ?? [];
             this.$nextTick(() => {
-              this.$refs.logScroller.scrollToPosition(Number.MAX_SAFE_INTEGER);
+              // the scroller is gone if the user left the page mid-request
+              this.$refs.logScroller?.scrollToPosition(Number.MAX_SAFE_INTEGER);
               this.isLoading = false;
             });
           })